// src/lib/authz/spicedb-generated-types.ts
export enum ERP_INVOICE_SUBJECT_RELATIONS {
  WORKSPACE_WORKSPACE = 'WORKSPACE_WORKSPACE',
}

export const ERP_INVOICE_SUBJECT_RELATIONS_MAP: SubjectRelationsMapType<
  RESOURCE_TYPES,
  ERP_INVOICE_RELATIONS,
  ERP_INVOICE_SUBJECT_RELATIONS
> = {
  WORKSPACE_WORKSPACE: {
    relation: ERP_INVOICE_RELATIONS.WORKSPACE,
    subjectType: RESOURCE_TYPES.ERP_WORKSPACE,
  },
} as const;

export enum ERP_INVOICE_SUBJECT_PERMISSIONS {
  USER_READ = 'USER_READ',
  USER_UPDATE = 'USER_UPDATE',
}
// ...
export enum ERP_WORKSPACE_SUBJECT_PERMISSIONS {
  USER_IS_ADMIN = 'USER_IS_ADMIN',
  USER_CAN_JOIN = 'USER_CAN_JOIN',
  USER_READ = 'USER_READ',
  USER_MANAGE = 'USER_MANAGE',
  USER_ADD_USER = 'USER_ADD_USER',
  USER_REMOVE_USER = 'USER_REMOVE_USER',
  USER_UPDATE_USER_ROLES = 'USER_UPDATE_USER_ROLES',
  USER_CAN_MANAGE_REFERENCE_NUMBER_TEMPLATES = 'USER_CAN_MANAGE_REFERENCE_NUMBER_TEMPLATES',
  USER_CAN_READ_REFERENCE_NUMBER_TEMPLATES = 'USER_CAN_READ_REFERENCE_NUMBER_TEMPLATES',
  USER_CAN_MANAGE_QUOTES = 'USER_CAN_MANAGE_QUOTES',
  USER_CAN_READ_QUOTES = 'USER_CAN_READ_QUOTES',
  USER_CAN_MANAGE_RFQS = 'USER_CAN_MANAGE_RFQS',
  USER_CAN_READ_RFQS = 'USER_CAN_READ_RFQS',
  USER_CAN_MANAGE_SALES_ORDERS = 'USER_CAN_MANAGE_SALES_ORDERS',
  USER_CAN_READ_SALES_ORDERS = 'USER_CAN_READ_SALES_ORDERS',
  USER_CAN_MANAGE_PURCHASE_ORDERS = 'USER_CAN_MANAGE_PURCHASE_ORDERS',
  USER_CAN_READ_PURCHASE_ORDERS = 'USER_CAN_READ_PURCHASE_ORDERS',
  USER_CAN_MANAGE_PRICE_BOOKS = 'USER_CAN_MANAGE_PRICE_BOOKS',
  USER_CAN_MANAGE_PRICES = 'USER_CAN_MANAGE_PRICES',
  USER_CREATE_PRICE_BOOK = 'USER_CREATE_PRICE_BOOK',
  USER_READ_PRICE_BOOKS = 'USER_READ_PRICE_BOOKS',
  USER_READ_PRICES = 'USER_READ_PRICES',
  USER_CAN_MANAGE_INTAKE_FORMS = 'USER_CAN_MANAGE_INTAKE_FORMS',
  USER_CREATE_INTAKE_FORM = 'USER_CREATE_INTAKE_FORM',
  USER_READ_INTAKE_FORMS = 'USER_READ_INTAKE_FORMS',
  USER_CAN_MANAGE_INTAKE_FORM_SUBMISSIONS = 'USER_CAN_MANAGE_INTAKE_FORM_SUBMISSIONS',
  USER_CREATE_INTAKE_FORM_SUBMISSION = 'USER_CREATE_INTAKE_FORM_SUBMISSION',
  USER_READ_INTAKE_FORM_SUBMISSIONS = 'USER_READ_INTAKE_FORM_SUBMISSIONS',
  USER_CAN_MANAGE_BUYER_INTAKE_FORM_SUBMISSIONS = 'USER_CAN_MANAGE_BUYER_INTAKE_FORM_SUBMISSIONS',
  USER_CAN_READ_BUYER_INTAKE_FORM_SUBMISSIONS = 'USER_CAN_READ_BUYER_INTAKE_FORM_SUBMISSIONS',
  USER_CAN_MANAGE_INVOICES = 'USER_CAN_MANAGE_INVOICES',
  USER_CAN_READ_INVOICES = 'USER_CAN_READ_INVOICES',
  USER_CAN_MANAGE_PROJECTS = 'USER_CAN_MANAGE_PROJECTS',
  USER_CAN_READ_PROJECTS = 'USER_CAN_READ_PROJECTS',
  USER_CAN_MANAGE_CONTACTS = 'USER_CAN_MANAGE_CONTACTS',
  USER_CAN_READ_CONTACTS = 'USER_CAN_READ_CONTACTS',
  USER_CAN_MANAGE_CHARGES = 'USER_CAN_MANAGE_CHARGES',
  USER_CAN_READ_CHARGES = 'USER_CAN_READ_CHARGES',
  USER_CAN_MANAGE_FILES = 'USER_CAN_MANAGE_FILES',
  USER_CAN_READ_FILES = 'USER_CAN_READ_FILES',
}